+++ fcsapi/test/adapter.test.ts
@@ -1,6 +1,7 @@
-const { assert } = require('chai')
-const { assertSuccess, assertError } = require('@chainlink/external-adapter')
-const { execute } = require('../adapter')
+import { assert } from 'chai'
+import { Requester, assertSuccess, assertError } from '@chainlink/external-adapter'
+import { AdapterRequest } from '@chainlink/types'
+import { execute } from '../src/adapter'
 
 describe('execute', () => {
   const jobID = '1'
@@ -21,13 +22,11 @@ describe('execute', () => {
     ]
 
     requests.forEach((req) => {
-      it(`${req.name}`, (done) => {
-        execute(req.testData, (statusCode, data) => {
-          assertSuccess({ expected: 200, actual: statusCode }, data, jobID)
-          assert.isAbove(Number(data.result), 0)
-          assert.isAbove(Number(data.data.result), 0)
-          done()
-        })
+      it(`${req.name}`, async () => {
+        const data = await execute(req.testData as AdapterRequest)
+        assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
+        assert.isAbove(data.result, 0)
+        assert.isAbove(data.data.result, 0)
       })
     })
   })
@@ -39,11 +38,13 @@ describe('execute', () => {
     ]
 
     requests.forEach((req) => {
-      it(`${req.name}`, (done) => {
-        execute(req.testData, (statusCode, data) => {
-          assertError({ expected: 400, actual: statusCode }, data, jobID)
-          done()
-        })
+      it(`${req.name}`, async () => {
+        try {
+          await execute(req.testData as AdapterRequest)
+        } catch (error) {
+          const errorResp = Requester.errored(jobID, error)
+          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
+        }
       })
     })
   })
@@ -57,11 +58,13 @@ describe('execute', () => {
     ]
 
     requests.forEach((req) => {
-      it(`${req.name}`, (done) => {
-        execute(req.testData, (statusCode, data) => {
-          assertError({ expected: 500, actual: statusCode }, data, jobID)
-          done()
-        })
+      it(`${req.name}`, async () => {
+        try {
+          await execute(req.testData as AdapterRequest)
+        } catch (error) {
+          const errorResp = Requester.errored(jobID, error)
+          assertError({ expected: 500, actual: errorResp.statusCode }, errorResp, jobID)
+        }
       })
     })
   })